fix(core): collect router metadata, split tests, add cnpmcore E2E
- AppWorkerLoader: collect router resolveModule in metadataOnly mode
  instead of skipping loadRouter entirely
- Fix agent?.close() for metadataOnly mode (agent not created)
- Split manifest.test.ts into 4 focused test files + shared helper
- Remove inline E2E tests from packages/core/test/
- Add ecosystem-ci/scripts/verify-manifest.mjs for E2E verification
- E2E workflow: generate manifest → boot cnpmcore → verify cache hits
