+++ packages/egg/src/lib/loader/AppWorkerLoader.ts
@@ -1,3 +1,5 @@
+import path from 'node:path';
+
 import { EggApplicationLoader } from './EggApplicationLoader.ts';
 
 /**
@@ -30,12 +32,6 @@ export class AppWorkerLoader extends EggApplicationLoader {
 
     // app > plugin
     await this.loadCustomApp();
-
-    // In metadataOnly mode, loadCustomApp triggers loadMetadata and marks ready.
-    // Skip the remaining phases (service/middleware/controller/router) since
-    // they do real module evaluation and are not needed for manifest generation.
-    if (this.options.metadataOnly) return;
-
     // app > plugin
     await this.loadService();
     // app > plugin > core
@@ -43,6 +39,11 @@ export class AppWorkerLoader extends EggApplicationLoader {
     // app
     await this.loadController();
     // app
-    await this.loadRouter(); // Depend on controllers
+    if (this.options.metadataOnly) {
+      // Resolve router path to collect metadata, but don't execute it
+      this.resolveModule(path.join(this.options.baseDir, 'app/router'));
+    } else {
+      await this.loadRouter(); // Depend on controllers
+    }
   }
 }
